+++ ui.frontend/html/src/components/StartSessionWizardView/ChooseFilesStep.jsx
@@ -1,6 +1,6 @@
 import { inject, observer } from 'mobx-react'
 import React from 'react'
-import TreeRoot from './FileTree/TreeRoot';
+import FileTree from './FileTree/FileTree';
 
 export default
 @inject('core', 'sessionUI')
@@ -8,7 +8,7 @@ export default
 class ChooseFilesStep extends React.Component {
   render() {
     return (
-      <TreeRoot
+      <FileTree
         roots={this.props.core.projectTrees.map(tree => tree.root)}
         checkedKeys={this.props.sessionUI.checkedKeys}
         onKeysChange={keys => this.props.sessionUI.setCheckedKeys(keys)}
