+++ ui.frontend/html/src/components/StartSessionWizardView/FileTree/FileTree.jsx
